+++ tests/test_sequence_feature_extraction_common.py
@@ -16,7 +16,7 @@
 
 import numpy as np
 
-from transformers import BatchSequenceFeature
+from transformers import BatchFeature
 from transformers.testing_utils import require_tf, require_torch
 
 from .test_feature_extraction_saving_common import FeatureExtractionSavingTestMixin
@@ -43,12 +43,12 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         feat_extract = self.feature_extraction_class(**self.feat_extract_dict)
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs})
+        processed_features = BatchFeature({input_name: speech_inputs})
 
         self.assertTrue(all(len(x) == len(y) for x, y in zip(speech_inputs, processed_features[input_name])))
 
         speech_inputs = self.feat_extract_tester.prepare_inputs_for_common(equal_length=True)
-        processed_features = BatchSequenceFeature({input_name: speech_inputs}, tensor_type="np")
+        processed_features = BatchFeature({input_name: speech_inputs}, tensor_type="np")
 
         batch_features_input = processed_features[input_name]
 
@@ -66,7 +66,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         feat_extract = self.feature_extraction_class(**self.feat_extract_dict)
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs}, tensor_type="pt")
+        processed_features = BatchFeature({input_name: speech_inputs}, tensor_type="pt")
 
         batch_features_input = processed_features[input_name]
 
@@ -84,7 +84,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         feat_extract = self.feature_extraction_class(**self.feat_extract_dict)
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs}, tensor_type="tf")
+        processed_features = BatchFeature({input_name: speech_inputs}, tensor_type="tf")
 
         batch_features_input = processed_features[input_name]
 
@@ -117,7 +117,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         speech_inputs = self.feat_extract_tester.prepare_inputs_for_common(numpify=numpify)
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs})
+        processed_features = BatchFeature({input_name: speech_inputs})
 
         pad_diff = self.feat_extract_tester.seq_length_diff
         pad_max_length = self.feat_extract_tester.max_seq_length + pad_diff
@@ -217,7 +217,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         speech_inputs = self.feat_extract_tester.prepare_inputs_for_common()
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs})
+        processed_features = BatchFeature({input_name: speech_inputs})
 
         input_np = feat_extract.pad(processed_features, padding="longest", return_tensors="np")[input_name]
         input_pt = feat_extract.pad(processed_features, padding="longest", return_tensors="pt")[input_name]
@@ -230,7 +230,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         speech_inputs = self.feat_extract_tester.prepare_inputs_for_common()
         input_name = feat_extract.model_input_names[0]
 
-        processed_features = BatchSequenceFeature({input_name: speech_inputs})
+        processed_features = BatchFeature({input_name: speech_inputs})
 
         input_np = feat_extract.pad(processed_features, padding="longest", return_tensors="np")[input_name]
         input_tf = feat_extract.pad(processed_features, padding="longest", return_tensors="tf")[input_name]
@@ -245,7 +245,7 @@ class SequenceFeatureExtractionTestMixin(FeatureExtractionSavingTestMixin):
         input_lenghts = [len(x) for x in speech_inputs]
         input_name = feat_extract.model_input_names[0]
 
-        processed = BatchSequenceFeature({input_name: speech_inputs})
+        processed = BatchFeature({input_name: speech_inputs})
 
         processed = feat_extract.pad(processed, padding="longest", return_tensors="np")
         self.assertIn("attention_mask", processed)
